requirements.txt
fastapi
uvicorn[standard]
google-genai
pillow
python-multipart
python-dotenv